refactor: lift worktree slug derivation into core::slug
The slug algorithm moves from hooks/template.rs into a pure core module so
the upcoming derived env values (daft env, #388) share one identity source
with the {worktree_slug} template variable. worktree_slug(ctx) stays as a
thin context adapter; no behavior change.
